+++ app/components/UI/EthInput/index.js
@@ -25,6 +25,7 @@ import { getTicker, getNormalizedTxState } from '../../../util/transactions';
 import FontAwesome from 'react-native-vector-icons/FontAwesome';
 import Device from '../../../util/Device';
 import NetworkMainAssetLogo from '../NetworkMainAssetLogo';
+import { toLowerCaseCompare } from '../../../util/general';
 
 const styles = StyleSheet.create({
 	root: {
@@ -282,8 +283,8 @@ class EthInput extends PureComponent {
 				});
 				break;
 			case 'CONTRACT_COLLECTIBLE_TRANSACTION': {
-				const collectiblesToShow = collectibles.filter(
-					collectible => collectible.address.toLowerCase() === transaction.selectedAsset.address.toLowerCase()
+				const collectiblesToShow = collectibles.filter(collectible =>
+					toLowerCaseCompare(collectible.address, transaction.selectedAsset.address)
 				);
 				this.setState({
 					assets: collectiblesToShow