+++ app/components/Views/SendFlow/Amount/index.js
@@ -27,7 +27,6 @@ import {
 	renderFromWei,
 	weiToFiat,
 	fromWei,
-	fromTokenMinimalUnit,
 	toWei,
 	isDecimal,
 	toTokenMinimalUnit,
@@ -36,7 +35,8 @@ import {
 	weiToFiatNumber,
 	balanceToFiatNumber,
 	getCurrencySymbol,
-	handleWeiNumber
+	handleWeiNumber,
+	fromTokenMinimalUnitString
 } from '../../../../util/number';
 import { getTicker, generateTransferData, getEther } from '../../../../util/transactions';
 import { util } from '@metamask/controllers';
@@ -53,6 +53,7 @@ import { ANALYTICS_EVENT_OPTS } from '../../../../util/analytics';
 import dismissKeyboard from 'react-native/Libraries/Utilities/dismissKeyboard';
 import NetworkMainAssetLogo from '../../../UI/NetworkMainAssetLogo';
 import { isMainNet } from '../../../../util/networks';
+import { toLowerCaseCompare } from '../../../../util/general';
 
 const { hexToBN, BNToHex } = util;
 
@@ -423,7 +424,7 @@ class Amount extends PureComponent {
 		} = this.props;
 		try {
 			const owner = await AssetsContractController.getOwnerOf(address, tokenId);
-			const isOwner = owner.toLowerCase() === selectedAddress.toLowerCase();
+			const isOwner = toLowerCaseCompare(owner, selectedAddress);
 			if (!isOwner) {
 				return strings('transaction.invalid_collectible_ownership');
 			}
@@ -656,10 +657,16 @@ class Amount extends PureComponent {
 		} else {
 			const exchangeRate = contractExchangeRates[selectedAsset.address];
 			if (internalPrimaryCurrencyIsCrypto || !exchangeRate) {
-				input = fromTokenMinimalUnit(contractBalances[selectedAsset.address], selectedAsset.decimals);
+				input = fromTokenMinimalUnitString(
+					contractBalances[selectedAsset.address]?.toString(10),
+					selectedAsset.decimals
+				);
 			} else {
 				input = `${balanceToFiatNumber(
-					fromTokenMinimalUnit(contractBalances[selectedAsset.address], selectedAsset.decimals),
+					fromTokenMinimalUnitString(
+						contractBalances[selectedAsset.address]?.toString(10),
+						selectedAsset.decimals
+					),
 					conversionRate,
 					exchangeRate
 				)}`;