+++ app/components/Views/SendFlow/Confirm/index.js
@@ -380,9 +380,14 @@ class Confirm extends PureComponent {
 		}
 	};
 
-	handleConfusables = async () => {
-		const { transactionToName } = this.props.transactionState;
-		await this.setState({ confusableCollection: collectConfusables(transactionToName) });
+	handleConfusables = () => {
+		const { identities = undefined, transactionState } = this.props;
+		const { transactionToName = undefined } = transactionState;
+		const accountNames = (identities && Object.keys(identities).map(hash => identities[hash].name)) || [];
+		const isOwnAccount = accountNames.includes(transactionToName);
+		if (transactionToName && !isOwnAccount) {
+			this.setState({ confusableCollection: collectConfusables(transactionToName) });
+		}
 	};
 
 	toggleWarningModal = () => this.setState(state => ({ warningModalVisible: !state.warningModalVisible }));
@@ -394,8 +399,8 @@ class Confirm extends PureComponent {
 		const { showCustomNonce, navigation, providerType } = this.props;
 		await this.handleFetchBasicEstimates();
 		showCustomNonce && (await this.setNetworkNonce());
-		await this.handleConfusables();
 		navigation.setParams({ providerType });
+		this.handleConfusables();
 		this.parseTransactionData();
 		this.prepareTransaction();
 	};